Remove minimum content size of 640x480 using setPreferredContentsSize
Because the QTWebPage viewport defaults to 640x480, contentSize will
never be smaller than the viewport. This means content smaller than this
is padded with the background white.

By setting the "preferred" content size to 1x1, we can remove that
minimum while still getting large content if explicit widths are set.
The downside here is that mobile optimized and responsive websites will
be rendered at their minimum widths. The best solution will probably be
to allow this to be configured.

Signed-off-by: Evan Owen <evan@cotap.com>